+++ src/Analysis/GWASV2/Components/Congratulations/Congratulations.css
@@ -1,0 +1,20 @@
+.configure-gwas_success {
+  margin-bottom: 60px;
+}
+.configure-gwas_success h3 {
+  text-align: center;
+}
+.configure-gwas_success button {
+  margin: 0 auto 0 auto;
+  width: 95%;
+  font-size: 14px;
+  color: #fff;
+  border-radius: 4px;
+  border: none;
+  background: #b35f05;
+  padding: 8px;
+  cursor: pointer;
+}
+.configure-gwas_success button:hover {
+  background: #784005;
+}
