+++ src/Analysis/GWASV2/Steps/ConfigureGWAS/ConfigureGWAS.jsx
@@ -7,6 +7,7 @@ import ACTIONS from '../../Utils/StateManagement/Actions';
 import DismissibleMessage from '../../Components/DismissibleMessage/DismissibleMessage';
 import { jobSubmission } from '../../Utils/gwasWorkflowApi';
 import { useSourceContext } from '../../Utils/Source';
+import Congratulations from '../../Components/Congratulations/Congratulations';
 import '../../GWASV2.css';
 import './ConfigureGWAS.css';
 
@@ -99,44 +100,12 @@ const ConfigureGWAS = ({
   return (
     <div className='configure-gwas'>
       {showSuccess && (
-        <div className='configure-gwas_success'>
-          <DismissibleMessage
-            title={`Congratulations on your submission for ${jobName}`}
-            description={`${successText}`}
-          />
-          <h3>DO YOU WANT TO</h3>
-          <div className='GWASUI-row'>
-            <div className='GWASUI-column'>
-              <a href='./GWASResults'>
-                <button type='button'>See Status</button>
-              </a>
-            </div>
-            <div className='GWASUI-column'>
-              <button
-                type='button'
-                onClick={() => {
-                  window.location.reload();
-                }}
-              >
-                Submit New Workflow
-              </button>
-            </div>
-            <div className='GWASUI-column'>
-              <button
-                type='button'
-                onClick={() => {
-                  setShowSuccess(false);
-                  dispatch({
-                    type: ACTIONS.SET_CURRENT_STEP,
-                    payload: 3,
-                  });
-                }}
-              >
-                Submit Similar (Stay Here)
-              </button>
-            </div>
-          </div>
-        </div>
+        <Congratulations
+          dispatch={dispatch}
+          setShowSuccess={setShowSuccess}
+          successText={successText}
+          jobName={jobName}
+        />
       )}
       {showError && (
         <DismissibleMessage
